This and that

+++ include/log.h
@@ -11,63 +11,63 @@
 
 extern _Bool EnableLogging;
 
-#define XLOG(...) if (EnableLogging) do { \
-                        GC_STOP(); \
-                        flockfile(stderr), \
-                        fprintf(stderr, "(%d) ", getpid()), \
-                        fprintf(stderr, "(%4llu) ", TID), \
-                        fprintf(stderr, __VA_ARGS__), \
-                        fprintf(stderr, "\n"), \
-                        funlockfile(stderr); \
-                        GC_RESUME(); \
+#define XLOG(...) if (EnableLogging) do {                    \
+                        GC_STOP();                           \
+                        flockfile(stderr),                   \
+                        fprintf(stderr, "(%d) ", getpid()),  \
+                        fprintf(stderr, "(%4lld) ", TID),    \
+                        fprintf(stderr, __VA_ARGS__),        \
+                        fprintf(stderr, "\n"),               \
+                        funlockfile(stderr);                 \
+                        GC_RESUME();                         \
                 } while (0)
 
-#define XLOGX(...) if (EnableLogging) do { \
-                        GC_STOP(); \
-                        flockfile(stderr), \
-                        fprintf(stderr, "(%d) ", getpid()), \
-                        fprintf(stderr, "(%4llu) ", TID), \
-                        fprintf(stderr, __VA_ARGS__), \
-                        funlockfile(stderr); \
-                        GC_RESUME(); \
+#define XLOGX(...) if (EnableLogging) do {                   \
+                        GC_STOP();                           \
+                        flockfile(stderr),                   \
+                        fprintf(stderr, "(%d) ", getpid()),  \
+                        fprintf(stderr, "(%4lld) ", TID),    \
+                        fprintf(stderr, __VA_ARGS__),        \
+                        funlockfile(stderr);                 \
+                        GC_RESUME();                         \
                 } while (0)
 
-#define XXLOG(...) if (EnableLogging) do { \
-                        Ty *ty = ty; \
-                        GC_STOP(); \
-                        flockfile(stderr), \
-                        fprintf(stderr, "(%d) ", getpid()), \
-                        fprintf(stderr, "(%4llu) ", TID), \
-                        fprintf(stderr, __VA_ARGS__), \
-                        fprintf(stderr, "\n"), \
-                        funlockfile(stderr); \
-                        GC_RESUME(); \
+#define XXLOG(...) if (EnableLogging) do {                   \
+                        Ty *ty = ty;                         \
+                        GC_STOP();                           \
+                        flockfile(stderr),                   \
+                        fprintf(stderr, "(%d) ", getpid()),  \
+                        fprintf(stderr, "(%4lld) ", TID),    \
+                        fprintf(stderr, __VA_ARGS__),        \
+                        fprintf(stderr, "\n"),               \
+                        funlockfile(stderr);                 \
+                        GC_RESUME();                         \
                 } while (0)
 
 #if !defined(TY_NO_LOG)
-#define LOG(...) if (EnableLogging) do { \
-                        GC_STOP(); \
-                        flockfile(stderr), \
-                        fprintf(stderr, "(%d) ", getpid()), \
-                        fprintf(stderr, "(%4llu) ", TID), \
-                        fprintf(stderr, __VA_ARGS__), \
-                        fprintf(stderr, "\n"), \
-                        funlockfile(stderr); \
-                        GC_RESUME(); \
+#define LOG(...) if (EnableLogging) do {                     \
+                        GC_STOP();                           \
+                        flockfile(stderr),                   \
+                        fprintf(stderr, "(%d) ", getpid()),  \
+                        fprintf(stderr, "(%4lld) ", TID),    \
+                        fprintf(stderr, __VA_ARGS__),        \
+                        fprintf(stderr, "\n"),               \
+                        funlockfile(stderr);                 \
+                        GC_RESUME();                         \
                 } while (0)
 #else
 #define LOG(...) ;
 #endif
 
-#define TID MyThreadId(ty)
+#define TID (long long)MyThreadId(ty)
 
 #if 0
-#define GCLOG(...) if (EnableLogging) do { \
-                        flockfile(stderr), \
-                        fprintf(stderr, "(%4llu) ", TID), \
-                        fprintf(stderr, __VA_ARGS__), \
-                        fprintf(stderr, "\n"), \
-                        funlockfile(stderr); \
+#define GCLOG(...) if (EnableLogging) do {                   \
+                        flockfile(stderr),                   \
+                        fprintf(stderr, "(%4lld) ", TID),    \
+                        fprintf(stderr, __VA_ARGS__),        \
+                        fprintf(stderr, "\n"),               \
+                        funlockfile(stderr);                 \
                 } while (0)
 #else
 #define GCLOG LOG
